Added Tileset Builder Window
The Tileset Builder Window is an Editor Window that takes in a set of sprites, and builds Tiles for each one, as well as a Tile Ruleset with all of the tiles.
You can then modify the tiles individually to change the adjacents. This window is made to just cut out the treadmill work of creating the tiles and dragging in the sprites.

Other additions:
- Added another test tileset
- Added another level texture

TODO: cleanup the code and make it check if the folders are valid before saving the assets to them (+ custom folder/tile ruleset name?)