Rebuilt Room Generation System
Room Generation is now done through ScriptableObjects - RoomGeneratorSO and GridObjectGeneratorSO.

The Room Generator stores a list of GridObject Generators, each accompanied by a colour. When generating a room, it goes through a texture and checks each texture to see if it matches one of the colours, and if so, makes the GridObjectGenerator create a GridObject for that space.

The Room Generator iterates through the grid after it has been generated, so that each GridObjectGenerator can build things based on where all the GridObjects are placed on the grid (e.g. dynamic tilesets).